fix(docs): Update documentation for HTML attribute elements in `Element` namespace.
