+++ tests/Element/HasAltTest.php
@@ -8,7 +8,7 @@ use PHPUnit\Framework\Attributes\{DataProviderExternal, Group};
 use PHPUnit\Framework\TestCase;
 use Stringable;
 use UIAwesome\Html\Attribute\Element\HasAlt;
-use UIAwesome\Html\Attribute\Tests\Support\Provider\Element\AltProvider;
+use UIAwesome\Html\Attribute\Tests\Provider\Element\AltProvider;
 use UIAwesome\Html\Attribute\Values\ElementAttribute;
 use UIAwesome\Html\Helper\Attributes;
 use UIAwesome\Html\Mixin\HasAttributes;
