+++ tests/Element/HasDecodingTest.php
@@ -9,7 +9,7 @@ use PHPUnit\Framework\Attributes\{DataProviderExternal, Group};
 use PHPUnit\Framework\TestCase;
 use Stringable;
 use UIAwesome\Html\Attribute\Element\HasDecoding;
-use UIAwesome\Html\Attribute\Tests\Support\Provider\Element\DecodingProvider;
+use UIAwesome\Html\Attribute\Tests\Provider\Element\DecodingProvider;
 use UIAwesome\Html\Attribute\Values\{Decoding, ElementAttribute};
 use UIAwesome\Html\Helper\{Attributes, Enum};
 use UIAwesome\Html\Helper\Exception\Message;
