+++ tests/Element/HasReferrerpolicyTest.php
@@ -9,7 +9,7 @@ use PHPUnit\Framework\Attributes\{DataProviderExternal, Group};
 use PHPUnit\Framework\TestCase;
 use Stringable;
 use UIAwesome\Html\Attribute\Element\HasReferrerpolicy;
-use UIAwesome\Html\Attribute\Tests\Support\Provider\Element\ReferrerpolicyProvider;
+use UIAwesome\Html\Attribute\Tests\Provider\Element\ReferrerpolicyProvider;
 use UIAwesome\Html\Attribute\Values\{ElementAttribute, Referrerpolicy};
 use UIAwesome\Html\Helper\{Attributes, Enum};
 use UIAwesome\Html\Helper\Exception\Message;
