+++ tests/Element/HasSrcTest.php
@@ -8,7 +8,7 @@ use PHPUnit\Framework\Attributes\{DataProviderExternal, Group};
 use PHPUnit\Framework\TestCase;
 use Stringable;
 use UIAwesome\Html\Attribute\Element\HasSrc;
-use UIAwesome\Html\Attribute\Tests\Support\Provider\Element\SrcProvider;
+use UIAwesome\Html\Attribute\Tests\Provider\Element\SrcProvider;
 use UIAwesome\Html\Attribute\Values\ElementAttribute;
 use UIAwesome\Html\Helper\Attributes;
 use UIAwesome\Html\Mixin\HasAttributes;
