+++ tests/Provider/Element/HrefProvider.php
@@ -2,21 +2,21 @@
 
 declare(strict_types=1);
 
-namespace UIAwesome\Html\Attribute\Tests\Support\Provider\Element;
+namespace UIAwesome\Html\Attribute\Tests\Provider\Element;
 
+use PHPForge\Support\Stub\{BackedString, Unit};
 use Stringable;
-use UIAwesome\Html\Attribute\Tests\Support\Stub\Values\{Backed, Unit};
 use UnitEnum;
 
 /**
- * Data provider for {@see \UIAwesome\Html\Attribute\Tests\Element\HasAltTest} test cases.
+ * Data provider for {@see \UIAwesome\Html\Attribute\Tests\Element\HasHrefTest} test cases.
  *
- * Provides representative input/output pairs for the `alt` attribute.
+ * Provides representative input/output pairs for the `href` attribute.
  *
- * @copyright Copyright (C) 2025 Terabytesoftw.
+ * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-final class AltProvider
+final class HrefProvider
 {
     /**
      * @phpstan-return array<
@@ -29,7 +29,7 @@ final class AltProvider
         $stringable = new class implements Stringable {
             public function __toString(): string
             {
-                return 'A descriptive alt text.';
+                return 'https://example.com/page';
             }
         };
 
@@ -41,18 +41,18 @@ final class AltProvider
                 '',
                 'Should return an empty string when setting an empty string.',
             ],
-            'enum backed' => [
-                Backed::VALUE,
+            'enum backed string' => [
+                BackedString::VALUE,
                 [],
-                Backed::VALUE,
-                ' alt="value"',
+                BackedString::VALUE,
+                ' href="value"',
                 'Should return the attribute value after setting it.',
             ],
             'enum unit' => [
                 Unit::value,
                 [],
                 Unit::value,
-                ' alt="value"',
+                ' href="value"',
                 'Should return the attribute value after setting it.',
             ],
             'null' => [
@@ -63,32 +63,46 @@ final class AltProvider
                 "Should return an empty string when the attribute is set to 'null'.",
             ],
             'replace existing' => [
-                'A descriptive alt text.',
-                ['alt' => 'A different alt text.'],
-                'A descriptive alt text.',
-                ' alt="A descriptive alt text."',
-                "Should return new 'alt' after replacing the existing 'alt' attribute.",
+                'https://example.com/new',
+                ['href' => 'https://example.com/old'],
+                'https://example.com/new',
+                ' href="https://example.com/new"',
+                "Should return new 'href' after replacing the existing 'href' attribute.",
             ],
             'string' => [
-                'A descriptive alt text.',
+                'https://example.com/page',
                 [],
-                'A descriptive alt text.',
-                ' alt="A descriptive alt text."',
+                'https://example.com/page',
+                ' href="https://example.com/page"',
+                'Should return the attribute value after setting it.',
+            ],
+            'string with fragment identifier' => [
+                '#section',
+                [],
+                '#section',
+                ' href="#section"',
+                'Should return the attribute value after setting it.',
+            ],
+            'string with relative path' => [
+                '/about',
+                [],
+                '/about',
+                ' href="/about"',
                 'Should return the attribute value after setting it.',
             ],
             'stringable' => [
                 $stringable,
                 [],
                 $stringable,
-                ' alt="A descriptive alt text."',
+                ' href="https://example.com/page"',
                 'Should return the attribute value after setting it with a Stringable instance.',
             ],
             'unset with null' => [
                 null,
-                ['alt' => 'A descriptive alt text.'],
+                ['href' => 'https://example.com/old'],
                 '',
                 '',
-                "Should unset the 'alt' attribute when 'null' is provided after a value.",
+                "Should unset the 'href' attribute when 'null' is provided after a value.",
             ],
         ];
     }
