+++ tests/Provider/Element/WidthProvider.php
@@ -2,26 +2,26 @@
 
 declare(strict_types=1);
 
-namespace UIAwesome\Html\Attribute\Tests\Support\Provider\Element;
+namespace UIAwesome\Html\Attribute\Tests\Provider\Element;
 
+use PHPForge\Support\Stub\{BackedString, Unit};
 use Stringable;
-use UIAwesome\Html\Attribute\Tests\Support\Stub\Values\{Backed, Unit};
 use UnitEnum;
 
 /**
- * Data provider for {@see \UIAwesome\Html\Attribute\Tests\Element\HasHeightTest} test cases.
+ * Data provider for {@see \UIAwesome\Html\Attribute\Tests\Element\HasWidthTest} test cases.
  *
- * Provides representative input/output pairs for the `height` attribute.
+ * Provides representative input/output pairs for the `width` attribute.
  *
  * @copyright Copyright (C) 2025 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-final class HeightProvider
+final class WidthProvider
 {
     /**
      * @phpstan-return array<
      *   string,
-     *   array{int|string|Stringable|UnitEnum|null, mixed[], int|string|Stringable|UnitEnum, string, string},
+     *   array{string|Stringable|UnitEnum|null, mixed[], string|Stringable|UnitEnum, string, string},
      * >
      */
     public static function values(): array
@@ -41,18 +41,18 @@ final class HeightProvider
                 '',
                 'Should return an empty string when setting an empty string.',
             ],
-            'enum backed' => [
-                Backed::VALUE,
+            'enum backed string' => [
+                BackedString::VALUE,
                 [],
-                Backed::VALUE,
-                ' height="value"',
+                BackedString::VALUE,
+                ' width="value"',
                 'Should return the attribute value after setting it.',
             ],
             'enum unit' => [
                 Unit::value,
                 [],
                 Unit::value,
-                ' height="value"',
+                ' width="value"',
                 'Should return the attribute value after setting it.',
             ],
             'null' => [
@@ -64,31 +64,31 @@ final class HeightProvider
             ],
             'replace existing' => [
                 '100px',
-                ['height' => '200px'],
+                ['width' => '200px'],
                 '100px',
-                ' height="100px"',
-                "Should return new 'height' after replacing the existing 'height' attribute.",
+                ' width="100px"',
+                "Should return new 'width' after replacing the existing 'width' attribute.",
             ],
             'string' => [
                 '100px',
                 [],
                 '100px',
-                ' height="100px"',
+                ' width="100px"',
                 'Should return the attribute value after setting it.',
             ],
             'stringable' => [
                 $stringable,
                 [],
                 $stringable,
-                ' height="100px"',
+                ' width="100px"',
                 'Should return the attribute value after setting it with a Stringable instance.',
             ],
             'unset with null' => [
                 null,
-                ['height' => '100px'],
+                ['width' => '100px'],
                 '',
                 '',
-                "Should unset the 'height' attribute when 'null' is provided after a value.",
+                "Should unset the 'width' attribute when 'null' is provided after a value.",
             ],
         ];
     }
